moving frame-related methods to new eta.core.frames module
