+++ eta/core/types.py
@@ -1009,7 +1009,7 @@ class VideoStreamInfo(JSONFile):
 class FrameRanges(JSONFile):
     '''A a monotonically increasing and disjoint series of frame ranges.
 
-    This type is implemented in ETA by the `eta.core.video.FrameRanges` class.
+    This type is implemented in ETA by the `eta.core.frames.FrameRanges` class.
 
     Examples:
         /path/to/frame_ranges.json
